+++ packages/addons/addons/magellan-shared/TransformationArguments.ts
@@ -6,5 +6,4 @@
  */
 export interface TransformationArguments {
     libPath: string;
-    functionsDir: string;
 }